+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Disable local variables when BetterErrors is detected.
+
+  *Joshua Wood*
+
 * Exit with 1 from deploy command when request fails but ignore the failures in
   the capistrano task.